move getUniqueConstraintsFields to shared
