+++ packages/twenty-shared/src/utils/indexMetadata/index.ts
@@ -1,0 +1,1 @@
+export * from './getUniqueConstraintsFields';
